ENH: moved SegmentationStep and EvaluationStep to separate files
