+++ SliceTracker/SliceTrackerUtils/steps/segmentation.py
@@ -1,0 +1,110 @@
+import os
+
+import EditorLib
+import qt
+import slicer
+import vtk
+from Editor import EditorWidget
+from base import SliceTrackerStepLogic, SliceTrackerStep
+from VolumeClipToLabel import VolumeClipToLabelWidget
+from SlicerProstateUtils.helpers import TargetCreationWidget
+
+
+class SliceTrackerSegmentationStepLogic(SliceTrackerStepLogic):
+
+  def __init__(self):
+    super(SliceTrackerSegmentationStepLogic, self).__init__()
+
+  def cleanup(self):
+    pass
+
+
+class SliceTrackerSegmentationStep(SliceTrackerStep):
+
+  NAME = "Segmentation"
+  LogicClass = SliceTrackerSegmentationStepLogic
+
+  def __init__(self):
+    self.modulePath = os.path.dirname(slicer.util.modulePath(self.MODULE_NAME)).replace(".py", "")
+    super(SliceTrackerSegmentationStep, self).__init__()
+
+  def setup(self):
+    self.setupIcons()
+    try:
+      import VolumeClipWithModel
+    except ImportError:
+      return slicer.util.warningDisplay("Error: Could not find extension VolumeClip. Open Slicer Extension Manager and "
+                                        "install VolumeClip.", "Missing Extension")
+
+    iconSize = qt.QSize(24, 24)
+
+    self.volumeClipGroupBox = qt.QWidget()
+    self.volumeClipGroupBoxLayout = qt.QVBoxLayout()
+    self.volumeClipGroupBox.setLayout(self.volumeClipGroupBoxLayout)
+
+    self.volumeClipToLabelWidget = VolumeClipToLabelWidget(self.volumeClipGroupBox)
+    self.volumeClipToLabelWidget.setup()
+    if qt.QSettings().value('Developer/DeveloperMode').lower() == 'true':
+      self.volumeClipToLabelWidget.reloadCollapsibleButton.hide()
+    self.volumeClipToLabelWidget.selectorsGroupBox.hide()
+    self.volumeClipToLabelWidget.colorGroupBox.hide()
+
+    # TODO!!!!
+    # self.volumeClipToLabelWidget.logic.colorNode = self.logic.mpReviewColorNode
+    # self.volumeClipToLabelWidget.onColorSelected(self.logic.segmentedLabelValue)
+
+    self.applyRegistrationButton = self.createButton("Apply Registration", icon=self.greenCheckIcon, iconSize=iconSize,
+                                                     toolTip="Run Registration.")
+    self.applyRegistrationButton.setFixedHeight(45)
+
+    self.editorWidgetButton = self.createButton("", icon=self.settingsIcon, toolTip="Show Label Editor",
+                                                enabled=False, iconSize=iconSize)
+
+    self.setupEditorWidget()
+    self.segmentationGroupBox = qt.QGroupBox()
+    self.segmentationGroupBoxLayout = qt.QGridLayout()
+    self.segmentationGroupBox.setLayout(self.segmentationGroupBoxLayout)
+    self.volumeClipToLabelWidget.segmentationButtons.layout().addWidget(self.editorWidgetButton)
+    self.segmentationGroupBoxLayout.addWidget(self.volumeClipGroupBox, 0, 0)
+    self.segmentationGroupBoxLayout.addWidget(self.editorWidgetParent, 1, 0)
+    self.segmentationGroupBoxLayout.addWidget(self.applyRegistrationButton, 2, 0)
+    self.segmentationGroupBoxLayout.setRowStretch(3, 1)
+    self.layout().addWidget(self.segmentationGroupBox)
+    self.editorWidgetParent.hide()
+
+    # TODO: control visibility of target settings table
+    self.setupTargetingStepUIElements()
+
+  def setupIcons(self):
+    self.greenCheckIcon = self.createIcon('icon-greenCheck.png')
+    self.settingsIcon = self.createIcon('icon-settings.png')
+
+  def setupEditorWidget(self):
+    self.editorWidgetParent = slicer.qMRMLWidget()
+    self.editorWidgetParent.setLayout(qt.QVBoxLayout())
+    self.editorWidgetParent.setMRMLScene(slicer.mrmlScene)
+    self.editUtil = EditorLib.EditUtil.EditUtil()
+    self.editorWidget = EditorWidget(parent=self.editorWidgetParent, showVolumesFrame=False)
+    self.editorWidget.setup()
+    self.editorParameterNode = self.editUtil.getParameterNode()
+
+  def setupTargetingStepUIElements(self):
+    self.targetingGroupBox = qt.QGroupBox()
+    self.targetingGroupBoxLayout = qt.QFormLayout()
+    self.targetingGroupBox.setLayout(self.targetingGroupBoxLayout)
+
+    self.fiducialsWidget = TargetCreationWidget(self.targetingGroupBoxLayout)
+    self.fiducialsWidget.addEventObserver(vtk.vtkCommand.ModifiedEvent, self.onTargetListModified)
+    self.finishTargetingStepButton = self.createButton("Done setting targets", enabled=True,
+                                                       toolTip="Click this button to continue after setting targets")
+
+    self.targetingGroupBoxLayout.addRow(self.finishTargetingStepButton)
+    self.layout().addWidget(self.targetingGroupBox)
+
+  def setupConnections(self):
+    super(SliceTrackerSegmentationStep, self).setupConnections()
+    self.fiducialsWidget.addEventObserver(vtk.vtkCommand.ModifiedEvent, self.onTargetListModified)
+
+  def onTargetListModified(self, caller, event):
+    self.finishTargetingStepButton.enabled = self.fiducialsWidget.currentNode is not None and \
+                                             self.fiducialsWidget.currentNode.GetNumberOfFiducials()
